requirements.txt
robotframework==5.0.1
robotframework-seleniumlibrary==6.0.0
robotframework-imaplibrary2==0.4.6
urllib3==1.26.14
requests==2.28.1
jsonpath-ng==1.5.3
jsonschema==4.17.0
pytz==2022.6
robotframework-databaselibrary==1.2.4
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.4
pymssql==2.2.7
selenium==4.5.0
robotframework-autoitlibrary==1.2.8
robotframework-excelutil==9.12
